+++ CHANGELOG.md
@@ -5,9 +5,19 @@ Version changes are pinned to SDK releases.
 
 ## Unreleased
 
-## [1.9.2] 2023-09-13
+## [1.10.1] 2023-09-15
 
 - New function to cancel all trigger orders + place order. ([#265](https://github.com/zetamarkets/sdk/pull/265))
+
+## [1.10.0] 2023-09-13
+
+- Subscribe to orderbooks using websockets instead of polling them. ([#262](https://github.com/zetamarkets/sdk/pull/262))
+
+## [1.9.2] 2023-09-12
+
+### Breaking
+
+- Add slot to callback functions. ([#270](https://github.com/zetamarkets/sdk/pull/270))
 
 ## [1.9.1] 2023-09-11